requirements-hf-embed.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements-hf-embed.txt requirements-hf-embed.in
#
aiohttp==3.9.5
    # via
    #   -c requirements.txt
    #   huggingface-hub
    #   llama-index-core
aiosignal==1.3.1
    # via
    #   -c requirements.txt
    #   aiohttp
annotated-types==0.7.0
    # via
    #   -c requirements.txt
    #   pydantic
anyio==4.4.0
    # via
    #   -c requirements.txt
    #   httpx
    #   openai
attrs==23.2.0
    # via
    #   -c requirements.txt
    #   aiohttp
certifi==2024.7.4
    # via
    #   -c requirements.txt
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via
    #   -c requirements.txt
    #   requests
click==8.1.7
    # via
    #   -c requirements.txt
    #   nltk
dataclasses-json==0.6.7
    # via llama-index-core
deprecated==1.2.14
    # via llama-index-core
dirtyjson==1.0.8
    # via llama-index-core
distro==1.9.0
    # via
    #   -c requirements.txt
    #   openai
filelock==3.15.4
    # via
    #   -c requirements.txt
    #   huggingface-hub
    #   torch
    #   transformers
frozenlist==1.4.1
    # via
    #   -c requirements.txt
    #   aiohttp
    #   aiosignal
fsspec==2024.6.1
    # via
    #   -c requirements.txt
    #   huggingface-hub
    #   llama-index-core
    #   torch
greenlet==3.0.3
    # via
    #   -c requirements.txt
    #   sqlalchemy
h11==0.14.0
    # via
    #   -c requirements.txt
    #   httpcore
httpcore==1.0.5
    # via
    #   -c requirements.txt
    #   httpx
httpx==0.27.0
    # via
    #   -c requirements.txt
    #   llama-cloud
    #   llama-index-core
    #   openai
huggingface-hub[inference]==0.23.4
    # via
    #   -c requirements.txt
    #   llama-index-embeddings-huggingface
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.7
    # via
    #   -c requirements.txt
    #   anyio
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.4
    # via
    #   -c requirements.txt
    #   torch
joblib==1.4.2
    # via
    #   nltk
    #   scikit-learn
llama-cloud==0.0.6
    # via llama-index-core
llama-index-core==0.10.52.post2
    # via
    #   -r requirements-hf-embed.in
    #   llama-index-embeddings-huggingface
llama-index-embeddings-huggingface==0.2.2
    # via -r requirements-hf-embed.in
markupsafe==2.1.5
    # via
    #   -c requirements.txt
    #   jinja2
marshmallow==3.21.3
    # via dataclasses-json
minijinja==2.0.1
    # via huggingface-hub
mpmath==1.3.0
    # via sympy
multidict==6.0.5
    # via
    #   -c requirements.txt
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
nest-asyncio==1.6.0
    # via llama-index-core
networkx==3.2.1
    # via
    #   -c requirements.txt
    #   llama-index-core
    #   torch
nltk==3.8.1
    # via llama-index-core
numpy==1.26.4
    # via
    #   -c requirements.txt
    #   llama-index-core
    #   pandas
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   transformers
openai==1.35.10
    # via
    #   -c requirements.txt
    #   llama-index-core
packaging==24.1
    # via
    #   -c requirements.txt
    #   huggingface-hub
    #   marshmallow
    #   transformers
pandas==2.2.2
    # via
    #   -c requirements.txt
    #   llama-index-core
pillow==10.4.0
    # via
    #   -c requirements.txt
    #   llama-index-core
    #   sentence-transformers
pydantic==2.8.2
    # via
    #   -c requirements.txt
    #   llama-cloud
    #   openai
pydantic-core==2.20.1
    # via
    #   -c requirements.txt
    #   pydantic
python-dateutil==2.9.0.post0
    # via
    #   -c requirements.txt
    #   pandas
pytz==2024.1
    # via
    #   -c requirements.txt
    #   pandas
pyyaml==6.0.1
    # via
    #   -c requirements.txt
    #   huggingface-hub
    #   llama-index-core
    #   transformers
regex==2024.5.15
    # via
    #   -c requirements.txt
    #   nltk
    #   tiktoken
    #   transformers
requests==2.32.3
    # via
    #   -c requirements.txt
    #   huggingface-hub
    #   llama-index-core
    #   tiktoken
    #   transformers
safetensors==0.4.3
    # via transformers
scikit-learn==1.5.1
    # via sentence-transformers
scipy==1.13.1
    # via
    #   -r requirements-hf-embed.in
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==3.0.1
    # via llama-index-embeddings-huggingface
six==1.16.0
    # via
    #   -c requirements.txt
    #   python-dateutil
sniffio==1.3.1
    # via
    #   -c requirements.txt
    #   anyio
    #   httpx
    #   openai
sqlalchemy[asyncio]==2.0.31
    # via
    #   llama-index-core
    #   sqlalchemy
sympy==1.13.0
    # via torch
tenacity==8.4.2
    # via
    #   -c requirements.txt
    #   llama-index-core
threadpoolctl==3.5.0
    # via scikit-learn
tiktoken==0.7.0
    # via
    #   -c requirements.txt
    #   llama-index-core
tokenizers==0.19.1
    # via
    #   -c requirements.txt
    #   transformers
torch==2.2.2
    # via sentence-transformers
tqdm==4.66.4
    # via
    #   -c requirements.txt
    #   huggingface-hub
    #   llama-index-core
    #   nltk
    #   openai
    #   sentence-transformers
    #   transformers
transformers==4.42.3
    # via sentence-transformers
typing-extensions==4.12.2
    # via
    #   -c requirements.txt
    #   huggingface-hub
    #   llama-index-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   torch
    #   typing-inspect
typing-inspect==0.9.0
    # via
    #   dataclasses-json
    #   llama-index-core
tzdata==2024.1
    # via
    #   -c requirements.txt
    #   pandas
urllib3==2.2.2
    # via
    #   -c requirements.txt
    #   requests
wrapt==1.16.0
    # via
    #   deprecated
    #   llama-index-core
yarl==1.9.4
    # via
    #   -c requirements.txt
    #   aiohttp
